fix float & button clr

+++ app/src/main/java/CalculatorOperations.java
@@ -1,17 +1,17 @@
 public class CalculatorOperations {
-    public static int tambah(int a, int b) {
+    public static double tambah(double a, double b) {
         return a + b;
     }
 
-    public static int kurang(int a, int b) {
+    public static double kurang(double a, double b) {
         return a - b;
     }
 
-    public static int kali(int a, int b) {
+    public static double kali(double a, double b) {
         return a * b;
     }
 
-    public static int bagi(int a, int b) {
+    public static double bagi(double a, double b) {
         return a / b; // Pastikan pembagi tidak nol
     }
 }
